+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- DeviceInputChangedListener component
 - Action Events into Sprite Tags
 
 ## [1.8.3] - 2026-03-20